requirements.txt
langchain
pandas
python-dotenv
tiktoken
jq
sentence-transformers
bs4
openai
pinecone-client
